[feat]: post 관련해서 자주 쓰는 함수 훅으로 빼서 사용
